+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_proof.test.ts
@@ -17,14 +17,14 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import { Web3Context } from 'web3-core';
 import { DEFAULT_RETURN_FORMAT, ETH_DATA_FORMAT, FMT_BYTES, FMT_NUMBER, format } from 'web3-utils';
 import { isNullish } from 'web3-validator';
+import { Web3EthExecutionAPI } from 'web3-types';
+import { ethRpcMethods } from 'web3-rpc-methods';
 
-import { getProof as rpcMethodsGetProof } from '../../../src/rpc_methods';
-import { Web3EthExecutionAPI } from '../../../src/web3_eth_execution_api';
 import { getProof } from '../../../src/rpc_method_wrappers';
 import { mockRpcResponse, testData } from './fixtures/get_proof';
 import { accountSchema } from '../../../src/schemas';
 
-jest.mock('../../../src/rpc_methods');
+jest.mock('web3-rpc-methods');
 
 describe('getProof', () => {
 	let web3Context: Web3Context<Web3EthExecutionAPI>;
@@ -54,7 +54,7 @@ describe('getProof', () => {
 			}
 
 			await getProof(web3Context, ...inputParameters, DEFAULT_RETURN_FORMAT);
-			expect(rpcMethodsGetProof).toHaveBeenCalledWith(
+			expect(ethRpcMethods.getProof).toHaveBeenCalledWith(
 				web3Context.requestManager,
 				inputAddress,
 				inputStorageKeyFormatted,
@@ -72,7 +72,7 @@ describe('getProof', () => {
 				mockRpcResponse,
 				expectedReturnFormat,
 			);
-			(rpcMethodsGetProof as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
+			(ethRpcMethods.getProof as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
 
 			const result = await getProof(web3Context, ...inputParameters, expectedReturnFormat);
 			expect(result).toStrictEqual(expectedFormattedResult);