+++ scripts/system_tests_utils.ts
@@ -119,6 +119,19 @@ export const closeOpenConnection = async (web3Context: Web3Context<any>) => {
 		await waitForOpenConnection(web3Context);
 	}
 
+	// If an error happened during closing, that is acceptable at tests, just print a 'warn'.
+	if (web3Context?.provider) {
+		(web3Context.provider as unknown as Web3BaseProvider).on('error', (err: any) => {
+			console.warn('error while trying to close the connection', err);
+		});
+	}
+
+	// Wait a bit to ensure the connection does not have a pending data that
+	//	could cause an error if written after closing the connection.
+	await new Promise<void>(resolve => {
+		setTimeout(resolve, 500);
+	});
+
 	if (
 		web3Context?.provider &&
 		'disconnect' in (web3Context.provider as unknown as Web3BaseProvider)